The outside track is deployed.
No errors, and it looks just "okay". Not very good, with some bugs, the frame rate of the tracking looks incompatable with the frame rate of the video. But at least, if one object is occluded or blocked, its bbox will be given by prediction. By the way, the frame rate after using this outside track methods seems lower than 12 fps with a mobile version of RTX 3050Ti GPU.

An additional file of simple statics of MOT 16/17 SOTA is also attached, it looks like the frame rate is not a big issue.