The method of outside tracks is deployed.
No compose errors. And the result can be seen on grafana dashboard. But the frame rate of tracking is different from the frame rate of the video. It looks like the tracking is slower, and the video will stuck after running for a while. (That might because I create too many tracklets).

By the way, a simple review of fps of some MOT 16/17 SOTA's is attached.