Fluency and flying bbox issues solved.
Now the tracker can run fluently and probably with no memory overflow with some techinical bugs (hard to explain in general).

There will be no flying bboxes with a lifeline managing mechanism for each tracklet.

It is ready (just ready) to include more trackers. Two trackers are already implemented in the code and they can run together, though they are not yet connected by the reasoner.